reorganize tokencache source
