+++ src/Microsoft.Identity.Client/Cache/NoLockTokenCacheProxy.cs
@@ -16,43 +16,32 @@ namespace Microsoft.Identity.Client.Cache
 
         }
 
-        public void Deserialize(byte[] msalV2State) => throw new NotImplementedException();
-
         public void DeserializeAdalV3(byte[] adalV3State)
         {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME
+#if !ANDROID && !iOS
             _tokenCache.DeserializeAdalV3NoLocks(adalV3State);
 #endif
         }
 
-        public void DeserializeMsalV2(byte[] msalV2State) => throw new NotImplementedException();
-
         public void DeserializeMsalV3(byte[] msalV3State)
         {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME
+#if !ANDROID && !iOS
             _tokenCache.DeserializeMsalV3NoLocks(msalV3State);
 #endif
         }
 
-        [Obsolete("This is expected to be removed in MSAL.NET v3 and ADAL.NET v5. We recommend using SerializeMsalV3/DeserializeMsalV3. Read more: https://aka.ms/msal-net-3x-cache-breaking-change", false)]
-        public void DeserializeUnifiedAndAdalCache(CacheData cacheData) => throw new NotImplementedException();
-
-        public byte[] Serialize() => throw new NotImplementedException();
-
         public byte[] SerializeAdalV3()
         {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME
+#if !ANDROID && !iOS
             return _tokenCache.SerializeAdalV3NoLocks();
 #else
             return null;
 #endif
         }
 
-        public byte[] SerializeMsalV2() => throw new NotImplementedException();
-
         public byte[] SerializeMsalV3()
         {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME
+#if !ANDROID && !iOS
             return _tokenCache.SerializeMsalV3NoLocks();
 #else
             return null;
@@ -60,13 +49,21 @@ namespace Microsoft.Identity.Client.Cache
         }
 
         [Obsolete("This is expected to be removed in MSAL.NET v3 and ADAL.NET v5. We recommend using SerializeMsalV3/DeserializeMsalV3. Read more: https://aka.ms/msal-net-3x-cache-breaking-change", false)]
-        public CacheData SerializeUnifiedAndAdalCache() => throw new NotImplementedException();
+        public void DeserializeUnifiedAndAdalCache(CacheData cacheData) => throw new NotImplementedException();
+
+        public byte[] Serialize() => throw new NotImplementedException(MsalErrorMessage.AkaMsmsalnet3BreakingChanges);
+        public void Deserialize(byte[] msalV2State) => throw new NotImplementedException(MsalErrorMessage.AkaMsmsalnet3BreakingChanges);
+        public byte[] SerializeMsalV2() => throw new NotImplementedException(MsalErrorMessage.AkaMsmsalnet3BreakingChanges);
+        public void DeserializeMsalV2(byte[] msalV2State) => throw new NotImplementedException();
+
+        [Obsolete("This is expected to be removed in MSAL.NET v3 and ADAL.NET v5. We recommend using SerializeMsalV3/DeserializeMsalV3. Read more: https://aka.ms/msal-net-3x-cache-breaking-change", false)]
+        public CacheData SerializeUnifiedAndAdalCache() => throw new NotImplementedException(MsalErrorMessage.AkaMsmsalnet3BreakingChanges);
 
         public void SetAfterAccess(TokenCacheCallback afterAccess) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
-        public void SetAsyncAfterAccess(Func<TokenCacheNotificationArgs, Task> afterAccess) => throw new NotImplementedException();
-        public void SetAsyncBeforeAccess(Func<TokenCacheNotificationArgs, Task> beforeAccess) => throw new NotImplementedException();
-        public void SetAsyncBeforeWrite(Func<TokenCacheNotificationArgs, Task> beforeWrite) => throw new NotImplementedException();
-        public void SetBeforeAccess(TokenCacheCallback beforeAccess) => throw new NotImplementedException();
-        public void SetBeforeWrite(TokenCacheCallback beforeWrite) => throw new NotImplementedException();
+        public void SetAsyncAfterAccess(Func<TokenCacheNotificationArgs, Task> afterAccess) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
+        public void SetAsyncBeforeAccess(Func<TokenCacheNotificationArgs, Task> beforeAccess) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
+        public void SetAsyncBeforeWrite(Func<TokenCacheNotificationArgs, Task> beforeWrite) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
+        public void SetBeforeAccess(TokenCacheCallback beforeAccess) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
+        public void SetBeforeWrite(TokenCacheCallback beforeWrite) => throw new NotImplementedException(MsalErrorMessage.FunctionalityNotAvailableInTokenCacheCallback);
     }
 }
